refactor(symbiosis): change file name
